# Parser: Multiple Return values & Producer Loop Improvements