… application/apps/indexer/Cargo.lock
… application/apps/indexer/Cargo.toml
… application/apps/indexer/addons/dlt-tools/Cargo.toml
… application/apps/indexer/indexer_cli/Cargo.toml
… application/apps/indexer/parsers/src/text.rs
… application/apps/indexer/processor/Cargo.toml
… application/apps/indexer/processor/src/export/mod.rs
… application/apps/indexer/session/src/handlers/export_raw.rs
… application/apps/indexer/session/src/handlers/observing/mod.rs
… application/apps/indexer/session/tests/snapshot_tests/snapshots/observe_dlt_with_someip_session.snap
… application/apps/indexer/sources/benches/mocks/mock_parser.rs
… application/apps/indexer/sources/src/command/process.rs
… application/apps/indexer/sources/src/producer.rs
… application/apps/indexer/sources/src/producer/tests/cancel_safety.rs
… application/apps/indexer/sources/src/producer/tests/mock_byte_source.rs
… application/apps/indexer/sources/src/producer/tests/mock_parser.rs
… application/apps/indexer/sources/src/producer/tests/multi_parse.rs
… application/apps/indexer/sources/src/producer/tests/single_parse.rs
… application/apps/indexer/sources/src/socket/mod.rs
… application/apps/indexer/sources/src/socket/tcp.rs
… application/apps/indexer/sources/src/socket/udp.rs
… application/apps/indexer/stypes/src/lib.rs
… application/apps/indexer/stypes/src/tests.rs
… application/apps/protocol/Cargo.lock
… application/apps/rustcore/rs-bindings/Cargo.lock
… application/apps/rustcore/ts-bindings/spec/session.observe.spec.ts
… cli/chipmunk-cli/Cargo.lock
… cli/chipmunk-cli/src/session/file.rs
… cli/chipmunk-cli/src/session/format/binary.rs
… cli/chipmunk-cli/src/session/format/mod.rs