This PR is a subset of #2244 without TCP reconnecting part

It provides the following changes:

### Parsers General:
* Make all parsers try to parse all possible items from the provided bytes instead of returning the first item only. If the first call of parse fails, then the error will be returned, however any errors after having a successful parse call will be ignored in the current call because parse will be called again with more data. If the error persists on the next call, then it will be returned directly since it will occur on the first parse call.
* Unit tests for all parsers adjusted to ensure the count of the parsed items on each call + New tests has been added.

### Producer Loop:

* Parsed items will be saved in an internal buffer reusing the memory to avoid costly memory allocation calls especially with parsers returning multiple items at one.
* Change producer API removing the stream interface and making `load_next_segment()` function returns a mutable reference of the newly loaded items, giving the caller the freedom to consume the items with `drain(..)` or iterate over their references.
* Current unit tests have been adjusted with the new producer API, and they all pass without any change in their core functionality. 
* Optimize handling parse error by avoiding dropping all loaded bytes. Instead, it will skip them one by one and reapply parse to avoid data loss.
* Abort the session on continues initial parse errors for the first megabyte of input bytes to avoid calling parse repeatedly on each byte drop of the source in case the input isn't suitable for the selected parser.
* Handling incomplete parse error is improved by Skipping one byte if load function didn't deliver any more bytes instead of terminating the whole session.
* Extend unit tests for producer with mocks to cover parse error cases.
* Extend unit tests for cancelation safety of producer loop covering all parser error cases.

### TCP/UDP Sources:
* Flushing the internal buffer is moved to `load()` method to tie freeing bytes in internal buffer and loading it with more data, making it more reliable to avoid potential bugs when having an incomplete parser result on last packet before having the internal buffer almost filled.

### SomeIP parser:
Fix converting errors from SomeIP crate into parser, ensuring `Incomplete` error type is delivered correctly in producer loop since this error will be handled different than parse errors.

## Performance Changes:
The changes led to slightly increasing in performance in spite of the parser returning multiple values. This improvement is thanks to reusing the memory with the internal buffer in producer loop. Here is a comparison between master and current PR.
```md
# Chipmunk App: Parsing DLT file 500MB:

### Current PR:
File Read Took: 12058 ms
File Read Took: 11972 ms
File Read Took: 12103 ms

### Master: 
File Read Took: 12648 ms
File Read Took: 12639 ms
File Read Took: 12680 ms

--------------------------------------------------

# Chipmunk App: Parsing DLT With attachments 150MB:

### Current PR:
File Read Took: 3509 ms
File Read Took: 3461 ms
File Read Took: 3538 ms

### Master: 
File Read Took: 3518 ms
File Read Took: 3485 ms
File Read Took: 3507 ms

--------------------------------------------------

# Rust Core Benchmarks: Mock parser returning 10 values on each call:

### Current PR:
time:   [5.9618 ms 5.9758 ms 5.9899 ms]
time:   [5.9676 ms 5.9816 ms 5.9954 ms]

### Master: 
time:   [8.3135 ms 8.3412 ms 8.3692 ms]
time:   [8.2700 ms 8.2920 ms 8.3141 ms]
```