@AmmarAbouZor 2025-03-14
Handling incomplete parse error is improved by Skipping one byte if load function didn't deliver any more bytes instead of terminating the whole session.